continue work on .env
