+++ packages/database/package.json
@@ -10,16 +10,18 @@
   },
   "scripts": {
     "init": "supabase login",
-    "dev": "supabase start",
+    "dev": "supabase start && tsx scripts/create_env.ts",
     "stop": "supabase stop",
     "check-types": "npm run lint && supabase stop && npm run dbdiff",
     "lint": "tsx scripts/lint.ts",
     "lint:fix": "tsx scripts/lint.ts -f",
     "build": "tsx scripts/build.ts",
     "test": "tsc && cucumber-js",
-    "genenv:supabasebranch": "tsx scripts/create_supabase_branch_env.ts",
-    "genenv:production": "vercel env pull --environment preview .env.production",
-    "genenv:local": "tsx scripts/create_supabase_local_env.ts",
+    "genenv": "tsx scripts/create_env.ts",
+    "genenv:branch": "tsx scripts/create_env.ts branch",
+    "genenv:production": "tsx scripts/create_env.ts production",
+    "genenv:local": "tsx scripts/create_env.ts local",
+    "genenv:all": "tsx scripts/create_env.ts all",
     "gentypes:production": "supabase start && supabase gen types typescript --project-id \"$SUPABASE_PROJECT_ID\" --schema public > types.gen.ts",
     "dbdiff": "supabase stop && supabase db diff",
     "dbdiff:save": "supabase stop && supabase db diff -f",
